FIX: Error control for really small datasets
We need samples labelled with all the different classes available for
each of the evaluation iterations we have. Really small datasets (e.g. less
than 10 samples) may not be able to provide all the different classes
available when iterating through the multiple evaluation runs we have.
We should skip those iterations as otherwise the provided data do not
match the tensor size otherwise.