tensorflow: reset graph in build_graph()
When we are building a new graph we want to clear the old one.

Without this, if the same process (e.g. a test) asks for a prediction
from a reloaded Classifier, it will fail with:

tensorflow.python.framework.errors_impl.NotFoundError: Key global-step_1 not found in checkpoint

and the old Clasifier leaks memory.