rework batch_size calculation, aiming to divide fairly
We had a hard batch size of 1000, which has curious effects when a
dataset contains a few more than a multiple of 1000; for example, if
there are 3001 rows, there would be three batchs of 1000 and one batch
of 1. With this new calculation we aim for 3 batches of 751, and one
batch of 749. We take care not to divide by zero with extreme values.

This also alters the n_epochs calculation, which had a special case
for small datasets.