webapp: debug option to stash incoming request data
If the MOODLE_ML_STASH_DIR is set, details of incoming requests are
stored in the specified directory. The data stored is sufficient to
replay the request in a test environment.

Forthcoming commits will introduce tools to inspect and analyse these
stashed requests, anonymise their contents so that derived data can be
publicly shared, and add tests based on this data.