Allow webapp to run without boto3 when not using s3
Users who are not using s3 should not have to install boto3.

while pylint and flake8 wisely do not currently complain about this
construct, it is easy to imagine a linter getting concerned in the
future, so we add a comment that people might read before they take
this out.