estimator: add balanced accuracy metric
Typical real world data for many problems will be skewed to one
outcome, which can make an intuitive understanding of accuracy scores
difficult.

For example, if a learning institution has a drop-out rate of 15%, the
classifier can get an accuracy score of 0.85 by saying students will
*never* drop out. 0.85 looks like a good score, but you could not say
this was a good classifier (or even, really, a classifier at all).

Another classifier that was 84% accurate in predicting both drop out
and finishing the course, its accuracy score would be 0.84, making it
look worse than the naive single-minded classifier, when really it is
better (as in, potentially useful in combination with other factors).

The balanced accuracy metric would give a score of 0.5 for the stupid
classifier, and 0.84 for the better classifier.