fix classifier export
The export function was not widely used or tested, so had bit-rotted
with the shift to Tensorflow 2 and Keras, and probably other changes.

The previous format was an ad-hoc json format that fails to capture all
the information for a successful re-import (it may have once worked, but
there are no tests to show that).

As we already can't import old models, we are able to change the export
format, and we have very good reasons for doing this:

1. The new format is close to industry conventions, using the Keras model
   save method for the core model, and pickle for metadata.

2. The new format is *exactly* what we use for internal saves. We know it
   works because it is used between every training run and the subsequent
   predictions.

3. Because import/export is the same as save/load other than file paths,
   we can avoid code duplication.