test_webapp: add degenerate_multiclass test
This is based on the Moodle PHP test that requires 100% accuracy for a
problem involving three examples with two inputs and three classes
(that is, one example per class). This is obviously learnable, but if
we don't test for it we can easily forget this case when setting
metaparameters and have a flapping test in Moodle.